app/requirements.txt
aiogram==3.1.1
google-generativeai==0.2.2
python-dotenv
aiosqlite

